+++ src/firmware/host/types/snp.rs
@@ -16,7 +16,7 @@ use crate::{
 };
 use std::{
     convert::{TryFrom, TryInto},
-    fmt::{self, Display, Formatter},
+    fmt::Display,
     io::{Read, Write},
     ops::BitOrAssign,
 };
@@ -25,7 +25,7 @@ use bitfield::bitfield;
 
 use self::FFI::types::SnpSetConfig;
 
-pub use crate::snp::types::TcbVersion;
+pub use crate::snp::types::{CertType, MaskId, TcbVersion};
 
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
@@ -46,136 +46,6 @@ bitfield! {
 impl BitOrAssign for SnpPlatformStatusFlags {
     fn bitor_assign(&mut self, rhs: Self) {
         self.0 |= rhs.0;
-    }
-}
-
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(Clone, Debug, Default, PartialEq, Eq)]
-#[repr(C)]
-/// Certificates which are accepted for [CertTableEntry](self::CertTableEntry)
-pub enum CertType {
-    /// Empty or closing entry for the CertTable
-    #[default]
-    Empty,
-
-    /// AMD Root Signing Key (ARK) certificate
-    ARK,
-
-    /// AMD SEV Signing Key (ASK) certificate
-    ASK,
-
-    /// Versioned Chip Endorsement Key (VCEK) certificate
-    VCEK,
-
-    /// Versioned Loaded Endorsement Key (VLEK) certificate
-    VLEK,
-
-    /// Certificate Revocation List (CRLs) certificate(s)
-    CRL,
-
-    /// Other (Specify GUID)
-    OTHER(uuid::Uuid),
-}
-
-impl Display for CertType {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        let guid = match self {
-            CertType::Empty => "00000000-0000-0000-0000-000000000000".to_string(),
-            CertType::ARK => "c0b406a4-a803-4952-9743-3fb6014cd0ae".to_string(),
-            CertType::ASK => "4ab7b379-bbac-4fe4-a02f-05aef327c782".to_string(),
-            CertType::VCEK => "63da758d-e664-4564-adc5-f4b93be8accd".to_string(),
-            CertType::VLEK => "a8074bc2-a25a-483e-aae6-39c045a0b8a1".to_string(),
-            CertType::CRL => "92f81bc3-5811-4d3d-97ff-d19f88dc67ea".to_string(),
-            CertType::OTHER(guid) => guid.to_string(),
-        };
-
-        write!(f, "{}", guid)
-    }
-}
-
-impl TryFrom<CertType> for uuid::Uuid {
-    type Error = uuid::Error;
-    fn try_from(value: CertType) -> Result<Self, Self::Error> {
-        match value {
-            CertType::Empty => uuid::Uuid::parse_str(&CertType::Empty.to_string()),
-            CertType::ARK => uuid::Uuid::parse_str(&CertType::ARK.to_string()),
-            CertType::ASK => uuid::Uuid::parse_str(&CertType::ASK.to_string()),
-            CertType::VCEK => uuid::Uuid::parse_str(&CertType::VCEK.to_string()),
-            CertType::VLEK => uuid::Uuid::parse_str(&CertType::VLEK.to_string()),
-            CertType::CRL => uuid::Uuid::parse_str(&CertType::CRL.to_string()),
-            CertType::OTHER(guid) => Ok(guid),
-        }
-    }
-}
-
-impl TryFrom<&uuid::Uuid> for CertType {
-    type Error = uuid::Error;
-
-    fn try_from(value: &uuid::Uuid) -> Result<Self, Self::Error> {
-        Ok(match value.to_string().as_str() {
-            "00000000-0000-0000-0000-000000000000" => CertType::Empty,
-            "c0b406a4-a803-4952-9743-3fb6014cd0ae" => CertType::ARK,
-            "4ab7b379-bbac-4fe4-a02f-05aef327c782" => CertType::ASK,
-            "63da758d-e664-4564-adc5-f4b93be8accd" => CertType::VCEK,
-            "a8074bc2-a25a-483e-aae6-39c045a0b8a1" => CertType::VLEK,
-            "92f81bc3-5811-4d3d-97ff-d19f88dc67ea" => CertType::CRL,
-            _ => CertType::OTHER(*value),
-        })
-    }
-}
-
-impl Ord for CertType {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        match (self, other) {
-            (Self::ARK, Self::ARK)
-            | (Self::ASK, Self::ASK)
-            | (Self::VCEK, Self::VCEK)
-            | (Self::VLEK, Self::VLEK)
-            | (Self::CRL, Self::CRL)
-            | (Self::Empty, Self::Empty) => std::cmp::Ordering::Equal,
-            (Self::OTHER(left), Self::OTHER(right)) => left.cmp(right),
-            (Self::Empty, _) => std::cmp::Ordering::Greater,
-            (_, Self::Empty) => std::cmp::Ordering::Less,
-            (Self::OTHER(_), _) => std::cmp::Ordering::Greater,
-            (_, Self::OTHER(_)) => std::cmp::Ordering::Less,
-            (Self::CRL, _) => std::cmp::Ordering::Greater,
-            (_, Self::CRL) => std::cmp::Ordering::Less,
-            (Self::ASK, _) => std::cmp::Ordering::Greater,
-            (_, Self::ASK) => std::cmp::Ordering::Less,
-            (Self::VLEK, _) => std::cmp::Ordering::Greater,
-            (_, Self::VLEK) => std::cmp::Ordering::Less,
-            (Self::VCEK, _) => std::cmp::Ordering::Greater,
-            (_, Self::VCEK) => std::cmp::Ordering::Less,
-        }
-    }
-}
-
-impl Encoder<()> for CertType {
-    fn encode(&self, writer: &mut impl Write, _: ()) -> Result<(), std::io::Error> {
-        let bytes = uuid::Uuid::try_from(self.clone())
-            .map_err(|_| std::io::ErrorKind::InvalidData)?
-            .into_bytes();
-        writer.write_bytes(bytes, ())?;
-        Ok(())
-    }
-}
-
-impl Decoder<()> for CertType {
-    fn decode(reader: &mut impl Read, _: ()) -> Result<Self, std::io::Error> {
-        let bytes = reader.read_bytes()?;
-        let uuid = uuid::Uuid::from_bytes(bytes);
-        Ok(CertType::try_from(&uuid).map_err(|_| std::io::ErrorKind::InvalidData)?)
-    }
-}
-
-impl ByteParser<()> for CertType {
-    type Bytes = [u8; 16];
-    const EXPECTED_LEN: Option<usize> = Some(16);
-}
-
-impl PartialOrd for CertType {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
     }
 }
 
@@ -272,16 +142,6 @@ impl PartialOrd for CertTableEntry {
     fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
         Some(self.cmp(other))
     }
-}
-
-/// Information regarding the SEV-SNP platform's TCB version.
-#[derive(Default, Clone, Debug, PartialEq, Eq)]
-pub struct TcbStatus {
-    /// Installed TCB version.
-    pub platform_version: TcbVersion,
-
-    /// Reported TCB version.
-    pub reported_version: TcbVersion,
 }
 
 bitfield! {
@@ -498,58 +358,6 @@ impl TryFrom<FFI::types::SnpSetConfig> for Config {
 }
 
 bitfield! {
-    /// Mask ID values that would go into an SNP CONFIG
-    ///
-    /// | Bit(s) | Name | Description |
-    /// |--------|------|-------------|
-    /// |0|MASK_CHIP_ID|Indicates that the CHIP_ID field in the attestation report will alwaysbe zero.|
-    /// |1|MASK_CHIP_KEY|Indicates that the VCEK is not used in attestation and guest key derivation.|
-    #[repr(C)]
-    #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-    #[derive(Default, Copy, Clone, PartialEq, Eq)]
-    pub struct MaskId(u32);
-    impl Debug;
-    /// Indicates that the CHIP_ID field in the attestation report will alwaysbe zero.
-    pub mask_chip_id, _: 0;
-    /// Indicates that the VCEK is not used in attestation and guest key derivation.
-    pub mask_chip_key, _: 1;
-}
-
-impl Encoder<()> for MaskId {
-    fn encode(&self, writer: &mut impl Write, _: ()) -> Result<(), std::io::Error> {
-        writer.write_bytes(self.0, ())?;
-        Ok(())
-    }
-}
-
-impl Decoder<()> for MaskId {
-    fn decode(reader: &mut impl Read, _: ()) -> Result<Self, std::io::Error> {
-        let mask = reader.read_bytes()?;
-        Ok(Self(mask))
-    }
-}
-
-impl ByteParser<()> for MaskId {
-    type Bytes = [u8; 4];
-    const EXPECTED_LEN: Option<usize> = Some(4);
-}
-
-impl Display for MaskId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            r#"
-    MaskID ({}):
-    Mask Chip ID: {}
-    ABI Chip Key: {}"#,
-            self.0,
-            self.mask_chip_id(),
-            self.mask_chip_key(),
-        )
-    }
-}
-
-bitfield! {
     /// Policy settings that appear in SNP PLATFORM STATUS
     ///
     /// | Bit(s) | Name | Description |
@@ -727,60 +535,6 @@ mod tests {
     use uuid::Uuid;
 
     #[test]
-    fn test_cert_type_sort_vcek() {
-        let mut certs: Vec<CertType> = vec![
-            CertType::Empty,
-            CertType::CRL,
-            CertType::OTHER(Uuid::parse_str("11111111-1111-1111-1111-111111111111").unwrap()),
-            CertType::OTHER(Uuid::parse_str("33333333-3333-3333-3333-333333333333").unwrap()),
-            CertType::OTHER(Uuid::parse_str("22222222-2222-2222-2222-222222222222").unwrap()),
-            CertType::ARK,
-            CertType::ASK,
-            CertType::VCEK,
-        ];
-
-        let sorted_certs: Vec<CertType> = vec![
-            CertType::ARK,
-            CertType::VCEK,
-            CertType::ASK,
-            CertType::CRL,
-            CertType::OTHER(Uuid::parse_str("11111111-1111-1111-1111-111111111111").unwrap()),
-            CertType::OTHER(Uuid::parse_str("22222222-2222-2222-2222-222222222222").unwrap()),
-            CertType::OTHER(Uuid::parse_str("33333333-3333-3333-3333-333333333333").unwrap()),
-            CertType::Empty,
-        ];
-        certs.sort();
-        assert_eq!(certs, sorted_certs);
-    }
-
-    #[test]
-    fn test_cert_type_sort_vlek() {
-        let mut certs: Vec<CertType> = vec![
-            CertType::Empty,
-            CertType::CRL,
-            CertType::OTHER(Uuid::parse_str("33333333-3333-3333-3333-333333333333").unwrap()),
-            CertType::OTHER(Uuid::parse_str("11111111-1111-1111-1111-111111111111").unwrap()),
-            CertType::OTHER(Uuid::parse_str("22222222-2222-2222-2222-222222222222").unwrap()),
-            CertType::ARK,
-            CertType::ASK,
-            CertType::VLEK,
-        ];
-
-        let sorted_certs: Vec<CertType> = vec![
-            CertType::ARK,
-            CertType::VLEK,
-            CertType::ASK,
-            CertType::CRL,
-            CertType::OTHER(Uuid::parse_str("11111111-1111-1111-1111-111111111111").unwrap()),
-            CertType::OTHER(Uuid::parse_str("22222222-2222-2222-2222-222222222222").unwrap()),
-            CertType::OTHER(Uuid::parse_str("33333333-3333-3333-3333-333333333333").unwrap()),
-            CertType::Empty,
-        ];
-        certs.sort();
-        assert_eq!(certs, sorted_certs);
-    }
-
-    #[test]
     fn test_snp_platform_status_flags_zeroed() {
         let actual: SnpPlatformStatusFlags = SnpPlatformStatusFlags(0);
 
@@ -796,39 +550,6 @@ mod tests {
         actual.0 |= 1 << 8;
         assert!(actual.is_owned());
         assert!(actual.is_encrypted_state_present());
-    }
-
-    #[test]
-    fn test_cert_type_fmt() {
-        let mut cert_type: CertType = CertType::Empty;
-        let mut expected: &str = "00000000-0000-0000-0000-000000000000";
-
-        assert_eq!(cert_type.to_string(), expected.to_string());
-
-        cert_type = CertType::ARK;
-        expected = "c0b406a4-a803-4952-9743-3fb6014cd0ae";
-
-        assert_eq!(cert_type.to_string(), expected.to_string());
-
-        cert_type = CertType::ASK;
-        expected = "4ab7b379-bbac-4fe4-a02f-05aef327c782";
-
-        assert_eq!(cert_type.to_string(), expected.to_string());
-
-        cert_type = CertType::VCEK;
-        expected = "63da758d-e664-4564-adc5-f4b93be8accd";
-
-        assert_eq!(cert_type.to_string(), expected.to_string());
-
-        cert_type = CertType::VLEK;
-        expected = "a8074bc2-a25a-483e-aae6-39c045a0b8a1";
-
-        assert_eq!(cert_type.to_string(), expected.to_string());
-
-        cert_type = CertType::CRL;
-        expected = "92f81bc3-5811-4d3d-97ff-d19f88dc67ea";
-
-        assert_eq!(cert_type.to_string(), expected.to_string());
     }
 
     #[test]
@@ -905,17 +626,6 @@ mod tests {
     }
 
     #[test]
-    fn test_cert_type_conversion() {
-        let ark_guid = Uuid::parse_str("c0b406a4-a803-4952-9743-3fb6014cd0ae").unwrap();
-        let cert_type = CertType::try_from(&ark_guid).unwrap();
-        assert_eq!(cert_type, CertType::ARK);
-
-        let uuid = Uuid::try_from(CertType::ARK).unwrap();
-        assert_eq!(uuid, ark_guid);
-    }
-
-    // Test Config struct and conversions
-    #[test]
     #[cfg(feature = "snp")]
     fn test_config() {
         let tcb = TcbVersion::new(None, 1, 2, 3, 4);
@@ -954,21 +664,6 @@ mod tests {
 
     // Test MaskId bitfield operations
     #[test]
-    fn test_mask_id() {
-        let mut mask = MaskId(0);
-        assert!(!mask.mask_chip_id());
-
-        mask.0 = 0x3;
-        assert!(mask.mask_chip_id());
-        assert!(mask.mask_chip_key());
-
-        // Test Display implementation
-        let display_output = format!("{}", mask);
-        assert!(display_output.contains("MaskID (3)"));
-    }
-
-    // Test SnpPlatformStatus
-    #[test]
     fn test_platform_status() {
         let status = SnpPlatformStatus::default();
         assert_eq!(status.state, 0);
@@ -982,22 +677,6 @@ mod tests {
     }
 
     // MaskId Tests
-    #[test]
-    fn test_mask_id_operations() {
-        let mut mask = MaskId(0);
-        assert!(!mask.mask_chip_id());
-        assert!(!mask.mask_chip_key());
-
-        mask.0 = 0x3;
-        assert!(mask.mask_chip_id());
-        assert!(mask.mask_chip_key());
-
-        let display = format!("{}", mask);
-        assert!(display.contains("MaskID (3)"));
-        assert!(display.contains("Mask Chip ID: true"));
-    }
-
-    // Config Tests
     #[test]
     #[cfg(feature = "snp")]
     fn test_config_conversions() {
@@ -1028,20 +707,6 @@ mod tests {
 
         status.platform_tcb_version = TcbVersion::new(None, 1, 2, 3, 4);
         assert_eq!(status.platform_tcb_version.snp, 3);
-    }
-
-    #[test]
-    fn test_tcb_status() {
-        let status = TcbStatus {
-            platform_version: TcbVersion::new(None, 1, 2, 3, 4),
-            reported_version: TcbVersion::new(None, 5, 6, 7, 8),
-        };
-
-        assert_eq!(status.platform_version.bootloader, 1);
-        assert_eq!(status.reported_version.bootloader, 5);
-
-        let default_status = TcbStatus::default();
-        assert_eq!(default_status.platform_version, Default::default());
     }
 
     #[test]
@@ -1159,17 +824,6 @@ mod tests {
     }
 
     #[test]
-    fn test_mask_id_boundary() {
-        let mut mask = MaskId(u32::MAX);
-        assert!(mask.mask_chip_id());
-        assert!(mask.mask_chip_key());
-
-        mask = MaskId(0);
-        assert!(!mask.mask_chip_id());
-        assert!(!mask.mask_chip_key());
-    }
-
-    #[test]
     fn test_config_reserved() {
         let config = Config::default();
         assert_eq!(config.reserved, [0u8; 52]);
@@ -1195,73 +849,6 @@ mod tests {
         assert!(status.platform_policy.ciphertext_hiding_dram_cap());
         assert!(status.platform_policy.ciphertext_hiding_dram_en());
         assert_eq!(status.state, 0xFF);
-    }
-
-    #[test]
-    fn test_cert_type_deserialization() {
-        let cert_types = vec![
-            CertType::Empty,
-            CertType::ARK,
-            CertType::ASK,
-            CertType::VCEK,
-            CertType::VLEK,
-            CertType::CRL,
-            CertType::OTHER(Uuid::parse_str("11111111-1111-1111-1111-111111111111").unwrap()),
-        ];
-
-        for cert_type in cert_types {
-            let serialized = cert_type.to_bytes().unwrap();
-            let deserialized = CertType::from_bytes(&serialized).unwrap();
-            assert_eq!(cert_type, deserialized);
-        }
-    }
-
-    #[test]
-    fn test_cert_type_try_from_uuid() {
-        // Test all valid UUIDs
-        let test_cases = vec![
-            ("00000000-0000-0000-0000-000000000000", CertType::Empty),
-            ("c0b406a4-a803-4952-9743-3fb6014cd0ae", CertType::ARK),
-            ("4ab7b379-bbac-4fe4-a02f-05aef327c782", CertType::ASK),
-            ("63da758d-e664-4564-adc5-f4b93be8accd", CertType::VCEK),
-            ("a8074bc2-a25a-483e-aae6-39c045a0b8a1", CertType::VLEK),
-            ("92f81bc3-5811-4d3d-97ff-d19f88dc67ea", CertType::CRL),
-            (
-                "11111111-1111-1111-1111-111111111111",
-                CertType::OTHER(Uuid::parse_str("11111111-1111-1111-1111-111111111111").unwrap()),
-            ),
-        ];
-
-        for (uuid_str, expected_type) in test_cases {
-            let uuid = Uuid::parse_str(uuid_str).unwrap();
-            assert_eq!(CertType::try_from(&uuid).unwrap(), expected_type);
-        }
-    }
-
-    #[test]
-    fn test_cert_type_cmp_complete() {
-        let mut cert_types = vec![
-            CertType::ARK,
-            CertType::VCEK,
-            CertType::VLEK,
-            CertType::ASK,
-            CertType::CRL,
-            CertType::Empty,
-            CertType::OTHER(Uuid::parse_str("11111111-1111-1111-1111-111111111111").unwrap()),
-        ];
-
-        let expected = vec![
-            CertType::ARK,
-            CertType::VCEK,
-            CertType::VLEK,
-            CertType::ASK,
-            CertType::CRL,
-            CertType::OTHER(Uuid::parse_str("11111111-1111-1111-1111-111111111111").unwrap()),
-            CertType::Empty,
-        ];
-
-        cert_types.sort();
-        assert_eq!(cert_types, expected);
     }
 
     #[test]
@@ -1296,24 +883,6 @@ mod tests {
         assert_eq!(sorted[3].cert_type, CertType::Empty);
     }
 
-    #[test]
-    fn test_mask_id_deserialization() {
-        let test_cases = vec![
-            MaskId(0),        // No bits set
-            MaskId(0x1),      // chip_id only
-            MaskId(0x2),      // chip_key only
-            MaskId(0x3),      // Both bits
-            MaskId(u32::MAX), // All bits
-        ];
-
-        for mask in test_cases {
-            let serialized = mask.clone().to_bytes().unwrap();
-            let deserialized = MaskId::from_bytes(&serialized).unwrap();
-            assert_eq!(mask.0, deserialized.0);
-            assert_eq!(mask.mask_chip_id(), deserialized.mask_chip_id());
-            assert_eq!(mask.mask_chip_key(), deserialized.mask_chip_key());
-        }
-    }
     #[test]
     fn test_cert_table_entry_complete_ordering() {
         let entries = vec![
@@ -1454,41 +1023,6 @@ mod tests {
     }
 
     #[test]
-    fn test_cert_type_to_uuid_conversion() {
-        use uuid::Uuid;
-
-        // Test successful conversions
-        assert_eq!(
-            Uuid::try_from(CertType::ARK).unwrap(),
-            Uuid::parse_str("c0b406a4-a803-4952-9743-3fb6014cd0ae").unwrap()
-        );
-        assert_eq!(
-            Uuid::try_from(CertType::ASK).unwrap(),
-            Uuid::parse_str("4ab7b379-bbac-4fe4-a02f-05aef327c782").unwrap()
-        );
-        assert_eq!(
-            Uuid::try_from(CertType::VCEK).unwrap(),
-            Uuid::parse_str("63da758d-e664-4564-adc5-f4b93be8accd").unwrap()
-        );
-        assert_eq!(
-            Uuid::try_from(CertType::VLEK).unwrap(),
-            Uuid::parse_str("a8074bc2-a25a-483e-aae6-39c045a0b8a1").unwrap()
-        );
-        assert_eq!(
-            Uuid::try_from(CertType::Empty).unwrap(),
-            Uuid::parse_str("00000000-0000-0000-0000-000000000000").unwrap()
-        );
-        assert_eq!(
-            Uuid::try_from(CertType::CRL).unwrap(),
-            Uuid::parse_str("92f81bc3-5811-4d3d-97ff-d19f88dc67ea").unwrap()
-        );
-        assert_eq!(
-            Uuid::try_from(CertType::OTHER(uuid::Uuid::max())).unwrap(),
-            Uuid::parse_str("ffffffff-ffff-ffff-ffff-ffffffffffff").unwrap()
-        );
-    }
-
-    #[test]
     #[cfg(target_os = "linux")]
     fn test_chain_visitor_methods() {
         // Test sequence visiting
@@ -1527,27 +1061,6 @@ mod tests {
         let init2 = PlatformInit(2);
         init1 |= init2;
         assert_eq!(init1.0, 3);
-    }
-
-    #[test]
-    fn test_mask_id_from_bytes() {
-        let bytes: [u8; 4] = [0b11, 0b11, 0b11, 0b11];
-        let mask_id = MaskId::from_bytes(&bytes).unwrap();
-        assert!(mask_id.mask_chip_id());
-        assert!(mask_id.mask_chip_key());
-    }
-
-    #[test]
-    fn test_mask_id_to_bytes() {
-        let mask_id = MaskId(0x01020304);
-        let bytes = mask_id.to_bytes().unwrap();
-        assert_eq!(bytes, [0x04, 0x03, 0x02, 0x01]);
-    }
-
-    #[test]
-    fn test_mask_id_default() {
-        let mask_id: MaskId = Default::default();
-        assert_eq!(mask_id.0, 0);
     }
 
     #[test]